Start the implementation of pipes flowing through the shapes and being turned on
Changed the data structures to use 2 bits instead of 1 for pipes, along
with a 4-type enumeration for pipe state.

Got the Hunt and Kill algorithm to make use of the new pipe states
correctly.

Rotating a piece will disable the pipes under its division of the tree.
Re-enabling pipes is not done yet.